Add better validation for the context that needs to be loaded
Since the context is not direcly loaded I was using a while loop with a
new promise. This way was working. However I was scared for infinite
loops.
I found out I could simply return out of the function when the context
is not loaded. And with useeffect dependency is on that particular
context it is able to reload when the context is available :)